// src/types/quantum.types.ts
export type ClassicalAlgorithm =
  | 'RSA-1024'
  | 'RSA-2048'
  | 'RSA-4096'
  | 'ECC-P256'
  | 'ECC-P384'
  | 'ECC-secp256k1'
  | 'ECDSA-P256'
  | 'ECDSA-P384'
  | 'DH-1024'
  | 'DH-2048'
  | 'DSA-1024'
  | 'DSA-2048'
  | 'PGP-RSA'
  | 'PGP-ECC'
  | 'S/MIME-RSA'
  | 'Ed25519'
  | 'X25519'
  | 'SHA-1'
  | 'SHA-256'
  | 'MD5'
  | 'TLS-1.0'
  | 'TLS-1.1'
  | 'TLS-1.2'
  | 'TLS-1.3'
  | 'AES-128'
  | 'AES-256'
  | '3DES'

export type QuantumThreatLevel = 'critical' | 'high' | 'medium' | 'low' | 'safe'

export type QuantumThreatType = 'shor' | 'grover' | 'both' | 'classical-broken' | 'safe'